Inital Commit
This is my inital commit of a python / django site. The blog and
photogallery base of it is started from the examples of this book:
Python Web Development with Django
Jeff Forcier, et al
The Stock Correlator is from the book:
Python Programming in Context
Bradley N. Miller David L. Ranum

I've incorporated the two examples to work together and started adding
functionality for a tool to help me in a game i've been playing recently
called Recetear.
Enjoy!